+++ CHANGELOG.md
@@ -85,6 +85,7 @@ December 13th, 2022
 - Added a configuration option to disable diagonal arrows.
     - When diagonal arrows are disabled, Assassin will point in increments of 90 degrees instead of 45.
 - Removed relay alert logic.
+- Added text to speech support.
 
 
 ## Version 2.0
@@ -99,3 +100,4 @@ This update adds support for optional internet based alerts to improve situation
 - Rearranged GPS configuration options.
     - All GPS settings are now in their own category.
 - Added basic GPS spoof detection system.
+- Moved text-to-speech configuration to the 'audio' section.